♻️ Move facetracking to own component
